feat: no erase zip options added for upload commands (#109)

+++ src/commands/commit.ts
@@ -77,7 +77,8 @@ export default new class Commit implements GluegunCommand {
 
     new RateLimit(apiRes.headers);
 
-    filesystem.remove(parameters.array[0]);
+    if (!parameters.options["no-erase-zip"])
+      filesystem.remove(parameters.array[0]);
 
     if (apiRes.status) {
       if (print.spinApiRes(apiRes, spin) > 399) return exit(apiRes.status);
