+++ src/commands/upload.ts
@@ -67,7 +67,9 @@ export default new class Upload implements GluegunCommand {
 
     new RateLimit(apiRes.headers);
 
-    filesystem.remove(parameters.array[0]);
+
+    if (!parameters.options["no-erase-zip"])
+      filesystem.remove(parameters.array[0]);
 
     if (apiRes.status) {
       if (print.spinApiRes(apiRes, spin) > 399) return exit(apiRes.status);
